Bob had some edits for Queen Victoria's Jubilee.
Start with an E attack note, so reformat the first part and add the
extra E for hte attack by moving the repeat bar and moving the B to the
end of the part. Point the first part ending D and b.

Add \grg to the 4-note groups in second part.

Adjust break to the jig.